make a widget and uses knockout to update instead of form

+++ ViewModels/UpdateShoppingCartItemViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class UpdateShoppingCartItemViewModel
     {
-        public decimal ProductId { get; set; }
+        public int ProductId { get; set; }
         public bool IsRemoved { get; set; }
         public int Quantity { get; set; }
     }
